Add tool to start local dped's via upstart
So this is a demo of how we could use upstart to
control the dpe processes. This just has a start.
It is really rough right now, but this should be better than
managing 100's of processes with waitid().

src/dped.conf needs to be in /etc/init/ and then
run "initictl reload-configuration".

Signed-off-by: Angus Salkeld <asalkeld@redhat.com>