Enable clone to determine platform/arch automatically so F14/F15/F16 work
This involves some significant rework to assembly.py and its handling
of constructors.  Also fixed various problems in this commit.  Something isn't
quite right with this commit - I think F14*jeos.tdl is being run multiple times
on an assembly create operaton.

The format of the database files have changed, requiring a rebuild of the JEOS
assembly and deployables.

Signed-off-by: Steven Dake <sdake@redhat.com>