CPE: support a configurable HTTP port to listen on
* src/common_agent.cpp: Replace the global program_name
variable with a proc_name member variable.
Add the "-H --http-port" command line option.
* src/common_agent.h (class CommonAgent): Store the http_port in this
base class, and provide a setter.  The getter is virtual, and
will return 0 unless overridden in a derived class.
"cpe" will override this to get the port, while "dpe" will
get 0, which will be used to distinguish each case in common code.
* cpe_agent.cpp (main): Construct the CpeHttpd object after
parsing the command line options, so that it may be constructed
with the appropriate port number.
* cpe_agent.h: Override the base class getter, to actually
return the assigned port number.
* cpe_httpd.h: Support construction with a specified port number.
* cpe_httpd.cpp: Likewise.
* init/pcloud-cped.service: Support reading options
from /etc/sysconfig/pacemaker-cloud

Signed-off-by: Pádraig Brady <P@draigBrady.com>