CPE: register the HTTP api with conductor
This uses libcurl to register our http interface with conductor.
We send the following to the conductor API:

  POST /api/hooks HTTP/1.1
  Content-type: application/xml
  Accept: application/xml

  <hook>
    <uri>http://localhost:8888/pacemaker-cloud/api</uri>
  </hook>

And in return get the following

  HTTP/1.1 201 Created
  Location: http://localhost:3000/api/hooks/1234
  Content-type: application/xml

  <hook id="1234" href="/api/hooks/1234">
    <uri>http://localhost:8888/pacemaker-cloud/api</uri>
  </hook>

* pacemaker-cloud.spec.in: Depend on libcurl
* src/Makefile.am (cped_SOURCES): Add the cpe_agent_register module.
(cped_LDFLAGS): Link with libcurl.
* src/cpe_agent.cpp (main): Register with conductor if
the cpe http interface started.
* src/cpe_agent.h (class CpeAgent): Add a variable to
record the conductor hook id url in case it's needed.
Declare the register_hook method.
* src/cpe_agent_register.cpp: A new module to use
libcurl to register with the conductor API.
Note the Location: is parsed from the HTTP header
currently, while the actual returned XML is not checked.

Signed-off-by: Pádraig Brady <P@draigBrady.com>