shell: fix loss of state for certain order of commands
If running multiple commands within a particular shell
instance, and one modifies assembly attributes
after the assembly has been added to a deployable,
then non default attributes of the assembly are lost.
For example, this sequence of commands would lose
the "openstack" attribute on the assemblies, thus
causing "libvirt" instances to be instantiated:

  deployable_create dep-wp openstack admin
  assembly_create assy-wordpress-F16 F16 x86_64
  deployable_assembly_add dep-wp assy-wordpress-F16
  assembly_resource_add http_wordpress httpd assy-wordpress-F16

The reason the last command causes state loss,
is because deployable.assembly_add() creates
a new assembly_factory.  This temporary factory
is then updated in vain by assembly_factory.register().
Instead we pass in the main assembly factory from pcloudsh.

Signed-off-by: Pádraig Brady <P@draigBrady.com>