Improve ta_disconnect and state machine to handle INVAL from wait_closed
libssh2_wait_closed could return LIBSSH2_ERROR_INVAL which indicates
there is more data to be read even though the channel is in a closed
state.  Now when this happens we execute a read operation to clear
the waiting packets until we can get a wait_closed operation to succeed.

Also as part of debugging a libssh2 leak, I have added a more complete
disconnect for ssh which frees all data associated with the ssh session.

The general idea here is to add a CTRL-C signal handler that can execute
a cleanup of the transport for assisting libssh2 upstream in debugging.

Signed-off-by: Steven Dake <sdake@redhat.com>