Add --enable-scale-bins configure options
The scale bins will produce a scale master which should be run once.
The scale dummy can be run hundreds of times and will read data from a file
stored by the scale master binary.  The file with the instance identifier is
stored in the current working directory and saved as the name of the assembly.

Signed-off-by: Steven Dake <sdake@redhat.com>