Add support for dynamic parameters (hostname and ipaddress).
So the point of this is to deal with cloud vms restarting with
a different ipaddress and not been able to deal with this in the
normal cluster way of setting up an extra ipaddress using the
IPaddress ocf resource (cloud infrastructure does not allow this).

So you define a simple binary (bash) and place a url into the
application xml within the services section.

cape will then setup an ocf resource (pacemaker-cloud/script_runner)
to run the user script. The ocf script_runner will have parameters
passed into it as environment variables.

See tests/check_reconfig.c for an example.

Signed-off-by: Angus Salkeld <asalkeld@redhat.com>